7.5-stable

+++ bin/cp/utils.c
@@ -1,4 +1,5 @@
-/*	$NetBSD: utils.c,v 1.4 1995/08/02 07:17:02 jtc Exp $	*/
+/*	$OpenBSD: utils.c,v 1.49 2021/10/24 21:24:21 deraadt Exp $	*/
+/*	$NetBSD: utils.c,v 1.6 1997/02/26 14:40:51 cgd Exp $	*/
 
 /*-
  * Copyright (c) 1991, 1993, 1994
@@ -12,11 +13,7 @@
  * 2. Redistributions in binary form must reproduce the above copyright
  *    notice, this list of conditions and the following disclaimer in the
  *    documentation and/or other materials provided with the distribution.
- * 3. All advertising materials mentioning features or use of this software
- *    must display the following acknowledgement:
- *	This product includes software developed by the University of
- *	California, Berkeley and its contributors.
- * 4. Neither the name of the University nor the names of its contributors
+ * 3. Neither the name of the University nor the names of its contributors
  *    may be used to endorse or promote products derived from this software
  *    without specific prior written permission.
  *
@@ -33,15 +30,7 @@
  * SUCH DAMAGE.
  */
 
-#ifndef lint
-#if 0
-static char sccsid[] = "@(#)utils.c	8.3 (Berkeley) 4/1/94";
-#else
-static char rcsid[] = "$NetBSD: utils.c,v 1.4 1995/08/02 07:17:02 jtc Exp $";
-#endif
-#endif /* not lint */
-
-#include <sys/param.h>
+#include <sys/types.h>
 #include <sys/stat.h>
 #include <sys/mman.h>
 #include <sys/time.h>
@@ -54,22 +43,38 @@ static char rcsid[] = "$NetBSD: utils.c,v 1.4 1995/08/02 07:17:02 jtc Exp $";
 #include <stdlib.h>
 #include <string.h>
 #include <unistd.h>
+#include <limits.h>
 
 #include "extern.h"
 
+#define _MAXBSIZE	(64 * 1024)
+
+int copy_overwrite(void);
+
 int
-copy_file(entp, dne)
-	FTSENT *entp;
-	int dne;
+copy_file(FTSENT *entp, int exists)
 {
-	static char buf[MAXBSIZE];
+	static char *buf;
+	static char *zeroes;
 	struct stat to_stat, *fs;
-	int ch, checkch, from_fd, rcount, rval, to_fd, wcount;
+	int from_fd, rcount, rval, to_fd, wcount;
+	const size_t buflen = _MAXBSIZE;
 #ifdef VM_AND_BUFFER_CACHE_SYNCHRONIZED
 	char *p;
 #endif
-	
-	if ((from_fd = open(entp->fts_path, O_RDONLY, 0)) == -1) {
+
+	if (!buf) {
+		buf = malloc(buflen);
+		if (!buf)
+			err(1, "malloc");
+	}
+	if (!zeroes) {
+		zeroes = calloc(1, buflen);
+		if (!zeroes)
+			err(1, "calloc");
+	}
+
+	if ((from_fd = open(entp->fts_path, O_RDONLY)) == -1) {
 		warn("%s", entp->fts_path);
 		return (1);
 	}
@@ -77,33 +82,33 @@ copy_file(entp, dne)
 	fs = entp->fts_statp;
 
 	/*
-	 * If the file exists and we're interactive, verify with the user.
+	 * In -f (force) mode, we always unlink the destination first
+	 * if it exists.  Note that -i and -f are mutually exclusive.
+	 */
+	if (exists && fflag)
+		(void)unlink(to.p_path);
+
+	/*
 	 * If the file DNE, set the mode to be the from file, minus setuid
 	 * bits, modified by the umask; arguably wrong, but it makes copying
 	 * executables work right and it's been that way forever.  (The
 	 * other choice is 666 or'ed with the execute bits on the from file
 	 * modified by the umask.)
 	 */
-	if (!dne) {
-		if (iflag) {
-			(void)fprintf(stderr, "overwrite %s? ", to.p_path);
-			checkch = ch = getchar();
-			while (ch != '\n' && ch != EOF)
-				ch = getchar();
-			if (checkch != 'y' && checkch != 'Y') {
-				(void)close(from_fd);
-				return (0);
-			}
-		}
-		to_fd = open(to.p_path, O_WRONLY | O_TRUNC, 0);
+	if (exists && !fflag) {
+		if (!copy_overwrite()) {
+			(void)close(from_fd);
+			return 2;
+ 		}
+		to_fd = open(to.p_path, O_WRONLY | O_TRUNC);
 	} else
 		to_fd = open(to.p_path, O_WRONLY | O_TRUNC | O_CREAT,
-		    fs->st_mode & ~(S_ISUID | S_ISGID));
+		    fs->st_mode & ~(S_ISTXT | S_ISUID | S_ISGID));
 
 	if (to_fd == -1) {
 		warn("%s", to.p_path);
 		(void)close(from_fd);
-		return (1);;
+		return (1);
 	}
 
 	rval = 0;
@@ -114,18 +119,20 @@ copy_file(entp, dne)
 	 * wins some CPU back.
 	 */
 #ifdef VM_AND_BUFFER_CACHE_SYNCHRONIZED
+	/* XXX broken for 0-size mmap */
 	if (fs->st_size <= 8 * 1048576) {
 		if ((p = mmap(NULL, (size_t)fs->st_size, PROT_READ,
-		    0, from_fd, (off_t)0)) == (char *)-1) {
-			warn("%s", entp->fts_path);
+		    MAP_FILE|MAP_SHARED, from_fd, (off_t)0)) == MAP_FAILED) {
+			warn("mmap: %s", entp->fts_path);
 			rval = 1;
 		} else {
+			madvise(p, fs->st_size, MADV_SEQUENTIAL);
 			if (write(to_fd, p, fs->st_size) != fs->st_size) {
 				warn("%s", to.p_path);
 				rval = 1;
 			}
 			/* Some systems don't unmap on close(2). */
-			if (munmap(p, fs->st_size) < 0) {
+			if (munmap(p, fs->st_size) == -1) {
 				warn("%s", entp->fts_path);
 				rval = 1;
 			}
@@ -133,15 +140,24 @@ copy_file(entp, dne)
 	} else
 #endif
 	{
-		while ((rcount = read(from_fd, buf, MAXBSIZE)) > 0) {
-			wcount = write(to_fd, buf, rcount);
+		int skipholes = 0;
+		struct stat tosb;
+		if (!fstat(to_fd, &tosb) && S_ISREG(tosb.st_mode))
+			skipholes = 1;
+		while ((rcount = read(from_fd, buf, buflen)) > 0) {
+			if (skipholes && memcmp(buf, zeroes, rcount) == 0)
+				wcount = lseek(to_fd, rcount, SEEK_CUR) == -1 ? -1 : rcount;
+			else
+				wcount = write(to_fd, buf, rcount);
 			if (rcount != wcount || wcount == -1) {
 				warn("%s", to.p_path);
 				rval = 1;
 				break;
 			}
 		}
-		if (rcount < 0) {
+		if (skipholes && rcount != -1)
+			rcount = ftruncate(to_fd, lseek(to_fd, 0, SEEK_CUR));
+		if (rcount == -1) {
 			warn("%s", entp->fts_path);
 			rval = 1;
 		}
@@ -161,7 +177,8 @@ copy_file(entp, dne)
 	 */
 #define	RETAINBITS \
 	(S_ISUID | S_ISGID | S_ISVTX | S_IRWXU | S_IRWXG | S_IRWXO)
-	else if (fs->st_mode & (S_ISUID | S_ISGID) && fs->st_uid == myuid)
+	if (!pflag && !exists &&
+	    fs->st_mode & (S_ISUID | S_ISGID) && fs->st_uid == myuid) {
 		if (fstat(to_fd, &to_stat)) {
 			warn("%s", to.p_path);
 			rval = 1;
@@ -170,6 +187,7 @@ copy_file(entp, dne)
 			warn("%s", to.p_path);
 			rval = 1;
 		}
+	}
 	(void)close(from_fd);
 	if (close(to_fd)) {
 		warn("%s", to.p_path);
@@ -179,34 +197,34 @@ copy_file(entp, dne)
 }
 
 int
-copy_link(p, exists)
-	FTSENT *p;
-	int exists;
+copy_link(FTSENT *p, int exists)
 {
 	int len;
-	char link[MAXPATHLEN];
+	char name[PATH_MAX];
 
-	if ((len = readlink(p->fts_path, link, sizeof(link))) == -1) {
+	if (exists && !copy_overwrite())
+		return (2);
+	if ((len = readlink(p->fts_path, name, sizeof(name)-1)) == -1) {
 		warn("readlink: %s", p->fts_path);
 		return (1);
 	}
-	link[len] = '\0';
+	name[len] = '\0';
 	if (exists && unlink(to.p_path)) {
 		warn("unlink: %s", to.p_path);
 		return (1);
 	}
-	if (symlink(link, to.p_path)) {
-		warn("symlink: %s", link);
+	if (symlink(name, to.p_path)) {
+		warn("symlink: %s", name);
 		return (1);
 	}
-	return (0);
+	return (pflag ? setfile(p->fts_statp, -1) : 0);
 }
 
 int
-copy_fifo(from_stat, exists)
-	struct stat *from_stat;
-	int exists;
+copy_fifo(struct stat *from_stat, int exists)
 {
+	if (exists && !copy_overwrite())
+		return (2);
 	if (exists && unlink(to.p_path)) {
 		warn("unlink: %s", to.p_path);
 		return (1);
@@ -215,14 +233,14 @@ copy_fifo(from_stat, exists)
 		warn("mkfifo: %s", to.p_path);
 		return (1);
 	}
-	return (pflag ? setfile(from_stat, 0) : 0);
+	return (pflag ? setfile(from_stat, -1) : 0);
 }
 
 int
-copy_special(from_stat, exists)
-	struct stat *from_stat;
-	int exists;
+copy_special(struct stat *from_stat, int exists)
 {
+	if (exists && !copy_overwrite())
+		return (2);
 	if (exists && unlink(to.p_path)) {
 		warn("unlink: %s", to.p_path);
 		return (1);
@@ -231,25 +249,42 @@ copy_special(from_stat, exists)
 		warn("mknod: %s", to.p_path);
 		return (1);
 	}
-	return (pflag ? setfile(from_stat, 0) : 0);
+	return (pflag ? setfile(from_stat, -1) : 0);
 }
 
+/*
+ * If the file exists and we're interactive, verify with the user.
+ */
+int
+copy_overwrite(void)
+{
+	int ch, checkch;
+
+	if (iflag) {
+		(void)fprintf(stderr, "overwrite %s? ", to.p_path);
+		checkch = ch = getchar();
+		while (ch != '\n' && ch != EOF)
+			ch = getchar();
+		if (checkch != 'y' && checkch != 'Y')
+			return (0);
+	}
+	return 1;
+}
 
 int
-setfile(fs, fd)
-	register struct stat *fs;
-	int fd;
+setfile(struct stat *fs, int fd)
 {
-	static struct timeval tv[2];
+	struct timespec ts[2];
 	int rval;
 
 	rval = 0;
-	fs->st_mode &= S_ISUID | S_ISGID | S_IRWXU | S_IRWXG | S_IRWXO;
+	fs->st_mode &= S_ISTXT | S_ISUID | S_ISGID | S_IRWXU | S_IRWXG | S_IRWXO;
 
-	TIMESPEC_TO_TIMEVAL(&tv[0], &fs->st_atimespec);
-	TIMESPEC_TO_TIMEVAL(&tv[1], &fs->st_mtimespec);
-	if (utimes(to.p_path, tv)) {
-		warn("utimes: %s", to.p_path);
+	ts[0] = fs->st_atim;
+	ts[1] = fs->st_mtim;
+	if (fd >= 0 ? futimens(fd, ts) :
+	    utimensat(AT_FDCWD, to.p_path, ts, AT_SYMLINK_NOFOLLOW)) {
+		warn("update times: %s", to.p_path);
 		rval = 1;
 	}
 	/*
@@ -258,32 +293,45 @@ setfile(fs, fd)
 	 * the mode; current BSD behavior is to remove all setuid bits on
 	 * chown.  If chown fails, lose setuid/setgid bits.
 	 */
-	if (fd ? fchown(fd, fs->st_uid, fs->st_gid) :
-	    chown(to.p_path, fs->st_uid, fs->st_gid)) {
+	if (fd >= 0 ? fchown(fd, fs->st_uid, fs->st_gid) :
+	    lchown(to.p_path, fs->st_uid, fs->st_gid)) {
 		if (errno != EPERM) {
 			warn("chown: %s", to.p_path);
 			rval = 1;
 		}
-		fs->st_mode &= ~(S_ISUID | S_ISGID);
+		fs->st_mode &= ~(S_ISTXT | S_ISUID | S_ISGID);
 	}
-	if (fd ? fchmod(fd, fs->st_mode) : chmod(to.p_path, fs->st_mode)) {
-		warn("chown: %s", to.p_path);
+	if (fd >= 0 ? fchmod(fd, fs->st_mode) :
+	    fchmodat(AT_FDCWD, to.p_path, fs->st_mode, AT_SYMLINK_NOFOLLOW)) {
+		warn("chmod: %s", to.p_path);
 		rval = 1;
 	}
 
-	if (fd ?
-	    fchflags(fd, fs->st_flags) : chflags(to.p_path, fs->st_flags)) {
-		warn("chflags: %s", to.p_path);
-		rval = 1;
-	}
+	/*
+	 * XXX
+	 * NFS doesn't support chflags; ignore errors unless there's reason
+	 * to believe we're losing bits.  (Note, this still won't be right
+	 * if the server supports flags and we were trying to *remove* flags
+	 * on a file that we copied, i.e., that we didn't create.)
+	 */
+	errno = 0;
+	if (fd >= 0 ? fchflags(fd, fs->st_flags) :
+	    chflagsat(AT_FDCWD, to.p_path, fs->st_flags, AT_SYMLINK_NOFOLLOW))
+		if (errno != EOPNOTSUPP || fs->st_flags != 0) {
+			warn("chflags: %s", to.p_path);
+			rval = 1;
+		}
 	return (rval);
 }
 
+
 void
-usage()
+usage(void)
 {
-	(void)fprintf(stderr, "%s\n%s\n",
-"usage: cp [-R [-H | -L | -P] [-fip] src target",
-"       cp [-R [-H | -L | -P] [-fip] src1 ... srcN directory");
+	(void)fprintf(stderr,
+	    "usage: %s [-afipv] [-R [-H | -L | -P]] source target\n", __progname);
+	(void)fprintf(stderr,
+	    "       %s [-afipv] [-R [-H | -L | -P]] source ... directory\n",
+	    __progname);
 	exit(1);
 }
